Move SliceArg trait and impls to slice.rs
Relocated the SliceArg trait and its implementations from base.rs to slice.rs for better code organization and separation of concerns. No functional changes were made.
